Refractor ExpressionBinderBase
