+++ src/Microsoft.AspNetCore.OData/Query/Expressions/ExpressionBinderBase.cs
@@ -121,1054 +121,6 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
             Model = model;
         }
 
-        [SuppressMessage("Microsoft.Maintainability", "CA1502:AvoidExcessiveComplexity", Justification = "These are simple conversion function and cannot be split up.")]
-        internal Expression CreateBinaryExpression(BinaryOperatorKind binaryOperator, Expression left, Expression right, bool liftToNull)
-        {
-            ExpressionType binaryExpressionType;
-
-            // When comparing an enum to a string, parse the string, convert both to the enum underlying type, and compare the values
-            // When comparing an enum to an enum with the same type, convert both to the underlying type, and compare the values
-            Type leftUnderlyingType = Nullable.GetUnderlyingType(left.Type) ?? left.Type;
-            Type rightUnderlyingType = Nullable.GetUnderlyingType(right.Type) ?? right.Type;
-
-            // Convert to integers unless Enum type is required
-            if ((TypeHelper.IsEnum(leftUnderlyingType) || TypeHelper.IsEnum(rightUnderlyingType)) && binaryOperator != BinaryOperatorKind.Has)
-            {
-                Type enumType = TypeHelper.IsEnum(leftUnderlyingType) ? leftUnderlyingType : rightUnderlyingType;
-                Type enumUnderlyingType = Enum.GetUnderlyingType(enumType);
-                left = ConvertToEnumUnderlyingType(left, enumType, enumUnderlyingType);
-                right = ConvertToEnumUnderlyingType(right, enumType, enumUnderlyingType);
-            }
-
-            if (leftUnderlyingType == typeof(DateTime) && rightUnderlyingType == typeof(DateTimeOffset))
-            {
-                right = DateTimeOffsetToDateTime(right, QuerySettings.TimeZone);
-            }
-            else if (rightUnderlyingType == typeof(DateTime) && leftUnderlyingType == typeof(DateTimeOffset))
-            {
-                left = DateTimeOffsetToDateTime(left, QuerySettings.TimeZone);
-            }
-
-            if ((IsDateOrOffset(leftUnderlyingType) && IsDate(rightUnderlyingType)) ||
-                (IsDate(leftUnderlyingType) && IsDateOrOffset(rightUnderlyingType)))
-            {
-                left = CreateDateBinaryExpression(left);
-                right = CreateDateBinaryExpression(right);
-            }
-
-            if ((IsDateOrOffset(leftUnderlyingType) && IsTimeOfDay(rightUnderlyingType)) ||
-                (IsTimeOfDay(leftUnderlyingType) && IsDateOrOffset(rightUnderlyingType)) ||
-                (IsTimeSpan(leftUnderlyingType) && IsTimeOfDay(rightUnderlyingType)) ||
-                (IsTimeOfDay(leftUnderlyingType) && IsTimeSpan(rightUnderlyingType)))
-            {
-                left = CreateTimeBinaryExpression(left);
-                right = CreateTimeBinaryExpression(right);
-            }
-
-            if (left.Type != right.Type)
-            {
-                // one of them must be nullable and the other is not.
-                left = ToNullable(left);
-                right = ToNullable(right);
-            }
-
-            if (left.Type == typeof(Guid) || right.Type == typeof(Guid))
-            {
-                switch (binaryOperator)
-                {
-                    case BinaryOperatorKind.GreaterThan:
-                    case BinaryOperatorKind.GreaterThanOrEqual:
-                    case BinaryOperatorKind.LessThan:
-                    case BinaryOperatorKind.LessThanOrEqual:
-                        left = Expression.Call(left, GuidCompareMethodInfo, right);
-                        right = ZeroConstant;
-                        break;
-                    default:
-                        break;
-                }
-            }
-
-            if (left.Type == typeof(string) || right.Type == typeof(string))
-            {
-                // convert nulls of type object to nulls of type string to make the String.Compare call work
-                left = ConvertNull(left, typeof(string));
-                right = ConvertNull(right, typeof(string));
-
-                // Use string.Compare instead of comparison for gt, ge, lt, le between two strings since direct comparisons are not supported
-                switch (binaryOperator)
-                {
-                    case BinaryOperatorKind.GreaterThan:
-                    case BinaryOperatorKind.GreaterThanOrEqual:
-                    case BinaryOperatorKind.LessThan:
-                    case BinaryOperatorKind.LessThanOrEqual:
-                        left = Expression.Call(StringCompareMethodInfo, left, right);
-                        right = ZeroConstant;
-                        break;
-                    default:
-                        break;
-                }
-            }
-
-            if (BinaryOperatorMapping.TryGetValue(binaryOperator, out binaryExpressionType))
-            {
-                if (left.Type == typeof(byte[]) || right.Type == typeof(byte[]))
-                {
-                    left = ConvertNull(left, typeof(byte[]));
-                    right = ConvertNull(right, typeof(byte[]));
-
-                    switch (binaryExpressionType)
-                    {
-                        case ExpressionType.Equal:
-                            return Expression.MakeBinary(binaryExpressionType, left, right, liftToNull, method: Linq2ObjectsComparisonMethods.AreByteArraysEqualMethodInfo);
-                        case ExpressionType.NotEqual:
-                            return Expression.MakeBinary(binaryExpressionType, left, right, liftToNull, method: Linq2ObjectsComparisonMethods.AreByteArraysNotEqualMethodInfo);
-                        default:
-                            IEdmPrimitiveType binaryType = typeof(byte[]).GetEdmPrimitiveType();
-                            throw new ODataException(Error.Format(SRResources.BinaryOperatorNotSupported, binaryType.FullName(), binaryType.FullName(), binaryOperator));
-                    }
-                }
-                else
-                {
-                    return Expression.MakeBinary(binaryExpressionType, left, right, liftToNull, method: null);
-                }
-            }
-            else
-            {
-                // Enum has a "has" operator
-                // {(c1, c2) => c1.HasFlag(Convert(c2))}
-                if (TypeHelper.IsEnum(left.Type) && TypeHelper.IsEnum(right.Type) && binaryOperator == BinaryOperatorKind.Has)
-                {
-                    UnaryExpression flag = Expression.Convert(right, typeof(Enum));
-                    return BindHas(left, flag);
-                }
-                else
-                {
-                    throw Error.NotSupported(SRResources.QueryNodeBindingNotSupported, binaryOperator, typeof(ExpressionBinderBase).Name);
-                }
-            }
-        }
-
-        internal Expression CreateConvertExpression(ConvertNode convertNode, Expression source)
-        {
-            Type conversionType = Model.GetClrType(convertNode.TypeReference, InternalAssembliesResolver);
-
-            if (conversionType == typeof(bool?) && source.Type == typeof(bool))
-            {
-                // we handle null propagation ourselves. So, if converting from bool to Nullable<bool> ignore.
-                return source;
-            }
-            else if (conversionType == typeof(Date?) &&
-                (source.Type == typeof(DateTimeOffset?) || source.Type == typeof(DateTime?)))
-            {
-                return source;
-            }
-            if ((conversionType == typeof(TimeOfDay?) && source.Type == typeof(TimeOfDay)) ||
-                ((conversionType == typeof(Date?) && source.Type == typeof(Date))))
-            {
-                return source;
-            }
-            else if (conversionType == typeof(TimeOfDay?) &&
-                (source.Type == typeof(DateTimeOffset?) || source.Type == typeof(DateTime?) || source.Type == typeof(TimeSpan?)))
-            {
-                return source;
-            }
-            else if (IsDateAndTimeRelated(conversionType) && IsDateAndTimeRelated(source.Type))
-            {
-                return source;
-            }
-            else if (source == NullConstant)
-            {
-                return source;
-            }
-            else
-            {
-                if (TypeHelper.IsEnum(source.Type))
-                {
-                    // we handle enum conversions ourselves
-                    return source;
-                }
-                else
-                {
-                    // if a cast is from Nullable<T> to Non-Nullable<T> we need to check if source is null
-                    if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True
-                        && IsNullable(source.Type) && !IsNullable(conversionType))
-                    {
-                        // source == null ? null : source.Value
-                        return
-                            Expression.Condition(
-                            test: CheckForNull(source),
-                            ifTrue: Expression.Constant(null, ToNullable(conversionType)),
-                            ifFalse: Expression.Convert(ExtractValueFromNullableExpression(source), ToNullable(conversionType)));
-                    }
-                    else
-                    {
-                        return Expression.Convert(source, conversionType);
-                    }
-                }
-            }
-        }
-
-
-        // If the expression is of non-standard edm primitive type (like uint), convert the expression to its standard edm type.
-        // Also, note that only expressions generated for ushort, uint and ulong can be understood by linq2sql and EF.
-        // The rest (char, char[], Binary) would cause issues with linq2sql and EF.
-        internal Expression ConvertNonStandardPrimitives(Expression source)
-        {
-            bool isNonstandardEdmPrimitive;
-            Type conversionType = source.Type.IsNonstandardEdmPrimitive(out isNonstandardEdmPrimitive);
-
-            if (isNonstandardEdmPrimitive)
-            {
-                Type sourceType = TypeHelper.GetUnderlyingTypeOrSelf(source.Type);
-
-                Contract.Assert(sourceType != conversionType);
-
-                Expression convertedExpression = null;
-
-                if (TypeHelper.IsEnum(sourceType))
-                {
-                    // we handle enum conversions ourselves
-                    convertedExpression = source;
-                }
-                else
-                {
-                    switch (Type.GetTypeCode(sourceType))
-                    {
-                        case TypeCode.UInt16:
-                        case TypeCode.UInt32:
-                        case TypeCode.UInt64:
-                            convertedExpression = Expression.Convert(ExtractValueFromNullableExpression(source), conversionType);
-                            break;
-
-                        case TypeCode.Char:
-                            convertedExpression = Expression.Call(ExtractValueFromNullableExpression(source), "ToString", typeArguments: null, arguments: null);
-                            break;
-
-                        case TypeCode.DateTime:
-                            convertedExpression = source;
-                            break;
-
-                        case TypeCode.Object:
-                            if (sourceType == typeof(char[]))
-                            {
-                                convertedExpression = Expression.New(typeof(string).GetConstructor(new[] { typeof(char[]) }), source);
-                            }
-                            else if (sourceType == typeof(XElement))
-                            {
-                                convertedExpression = Expression.Call(source, "ToString", typeArguments: null, arguments: null);
-                            }
-#if NETFX // System.Data.Linq.Binary is only supported in the AspNet version.
-                            else if (sourceType == typeof(Binary))
-                            {
-                                convertedExpression = Expression.Call(source, "ToArray", typeArguments: null, arguments: null);
-                            }
-#endif
-                            break;
-
-                        default:
-                            Contract.Assert(false, Error.Format("missing non-standard type support for {0}", sourceType.Name));
-                            break;
-                    }
-                }
-
-                if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True && IsNullable(source.Type))
-                {
-                    // source == null ? null : source
-                    return Expression.Condition(
-                        CheckForNull(source),
-                        ifTrue: Expression.Constant(null, ToNullable(convertedExpression.Type)),
-                        ifFalse: ToNullable(convertedExpression));
-                }
-                else
-                {
-                    return convertedExpression;
-                }
-            }
-
-            return source;
-        }
-
-        internal Expression MakePropertyAccess(PropertyInfo propertyInfo, Expression argument)
-        {
-            Expression propertyArgument = argument;
-            if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True)
-            {
-                // we don't have to check if the argument is null inside the function call as we do it already
-                // before calling the function. So remove the redundant null checks.
-                propertyArgument = RemoveInnerNullPropagation(argument);
-            }
-
-            // if the argument is of type Nullable<T>, then translate the argument to Nullable<T>.Value as none
-            // of the canonical functions have overloads for Nullable<> arguments.
-            propertyArgument = ExtractValueFromNullableExpression(propertyArgument);
-
-            return Expression.Property(propertyArgument, propertyInfo);
-        }
-
-        // creates an expression for the corresponding OData function.
-        internal Expression MakeFunctionCall(MemberInfo member, params Expression[] arguments)
-        {
-            Contract.Assert(member.MemberType == MemberTypes.Property || member.MemberType == MemberTypes.Method);
-
-            IEnumerable<Expression> functionCallArguments = arguments;
-            if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True)
-            {
-                // we don't have to check if the argument is null inside the function call as we do it already
-                // before calling the function. So remove the redundant null checks.
-                functionCallArguments = arguments.Select(a => RemoveInnerNullPropagation(a));
-            }
-
-            // if the argument is of type Nullable<T>, then translate the argument to Nullable<T>.Value as none
-            // of the canonical functions have overloads for Nullable<> arguments.
-            functionCallArguments = ExtractValueFromNullableArguments(functionCallArguments);
-
-            Expression functionCall;
-            if (member.MemberType == MemberTypes.Method)
-            {
-                MethodInfo method = member as MethodInfo;
-                if (method.IsStatic)
-                {
-                    functionCall = Expression.Call(null, method, functionCallArguments);
-                }
-                else
-                {
-                    functionCall = Expression.Call(functionCallArguments.First(), method, functionCallArguments.Skip(1));
-                }
-            }
-            else
-            {
-                // property
-                functionCall = Expression.Property(functionCallArguments.First(), member as PropertyInfo);
-            }
-
-            return CreateFunctionCallWithNullPropagation(functionCall, arguments);
-        }
-
-        internal Expression CreateFunctionCallWithNullPropagation(Expression functionCall, Expression[] arguments)
-        {
-            if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True)
-            {
-                Expression test = CheckIfArgumentsAreNull(arguments);
-
-                if (test == FalseConstant)
-                {
-                    // none of the arguments are/can be null.
-                    // so no need to do any null propagation
-                    return functionCall;
-                }
-                else
-                {
-                    // if one of the arguments is null, result is null (not defined)
-                    return
-                        Expression.Condition(
-                        test: test,
-                        ifTrue: Expression.Constant(null, ToNullable(functionCall.Type)),
-                        ifFalse: ToNullable(functionCall));
-                }
-            }
-            else
-            {
-                return functionCall;
-            }
-        }
-
-        // we don't have to do null checks inside the function for arguments as we do the null checks before calling
-        // the function when null propagation is enabled.
-        // this method converts back "arg == null ? null : convert(arg)" to "arg"
-        // Also, note that we can do this generically only because none of the odata functions that we support can take null
-        // as an argument.
-        internal Expression RemoveInnerNullPropagation(Expression expression)
-        {
-            Contract.Assert(expression != null);
-
-            if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True)
-            {
-                // only null propagation generates conditional expressions
-                if (expression.NodeType == ExpressionType.Conditional)
-                {
-                    // make sure to skip the DateTime IFF clause
-                    ConditionalExpression conditionalExpr = (ConditionalExpression)expression;
-                    if (conditionalExpr.Test.NodeType != ExpressionType.OrElse)
-                    {
-                        expression = conditionalExpr.IfFalse;
-                        Contract.Assert(expression != null);
-
-                        if (expression.NodeType == ExpressionType.Convert)
-                        {
-                            UnaryExpression unaryExpression = expression as UnaryExpression;
-                            Contract.Assert(unaryExpression != null);
-
-                            if (Nullable.GetUnderlyingType(unaryExpression.Type) == unaryExpression.Operand.Type)
-                            {
-                                // this is a cast from T to Nullable<T> which is redundant.
-                                expression = unaryExpression.Operand;
-                            }
-                        }
-                    }
-                }
-            }
-
-            return expression;
-        }
-
-        internal string GetFullPropertyPath(SingleValueNode node)
-        {
-            string path = null;
-            SingleValueNode parent = null;
-            switch (node.Kind)
-            {
-                case QueryNodeKind.SingleComplexNode:
-                    var complexNode = (SingleComplexNode)node;
-                    path = complexNode.Property.Name;
-                    parent = complexNode.Source;
-                    break;
-                case QueryNodeKind.SingleValuePropertyAccess:
-                    var propertyNode = ((SingleValuePropertyAccessNode)node);
-                    path = propertyNode.Property.Name;
-                    parent = propertyNode.Source;
-                    break;
-                case QueryNodeKind.SingleNavigationNode:
-                    var navNode = ((SingleNavigationNode)node);
-                    path = navNode.NavigationProperty.Name;
-                    parent = navNode.Source;
-                    break;
-            }
-
-            if (parent != null)
-            {
-                var parentPath = GetFullPropertyPath(parent);
-                if (parentPath != null)
-                {
-                    path = parentPath + "\\" + path;
-                }
-            }
-
-            return path;
-        }
-
-        /// <summary>
-        /// Gets property for dynamic properties dictionary.
-        /// </summary>
-        /// <param name="openNode"></param>
-        /// <returns>Returns CLR property for dynamic properties container.</returns>
-        protected PropertyInfo GetDynamicPropertyContainer(SingleValueOpenPropertyAccessNode openNode)
-        {
-            if (openNode == null)
-            {
-                throw Error.ArgumentNull(nameof(openNode));
-            }
-
-            IEdmStructuredType edmStructuredType;
-            IEdmTypeReference edmTypeReference = openNode.Source.TypeReference;
-            if (edmTypeReference.IsEntity())
-            {
-                edmStructuredType = edmTypeReference.AsEntity().EntityDefinition();
-            }
-            else if (edmTypeReference.IsComplex())
-            {
-                edmStructuredType = edmTypeReference.AsComplex().ComplexDefinition();
-            }
-            else
-            {
-                throw Error.NotSupported(SRResources.QueryNodeBindingNotSupported, openNode.Kind, typeof(ExpressionBinderBase).Name);
-            }
-
-            return Model.GetDynamicPropertyDictionary(edmStructuredType);
-        }
-
-        private static Expression CheckIfArgumentsAreNull(Expression[] arguments)
-        {
-            if (arguments.Any(arg => arg == NullConstant))
-            {
-                return TrueConstant;
-            }
-
-            arguments =
-                arguments
-                .Select(arg => CheckForNull(arg))
-                .Where(arg => arg != null)
-                .ToArray();
-
-            if (arguments.Any())
-            {
-                return arguments
-                    .Aggregate((left, right) => Expression.OrElse(left, right));
-            }
-            else
-            {
-                return FalseConstant;
-            }
-        }
-
-        internal static Expression CheckForNull(Expression expression)
-        {
-            if (IsNullable(expression.Type) && expression.NodeType != ExpressionType.Constant)
-            {
-                return Expression.Equal(expression, Expression.Constant(null));
-            }
-            else
-            {
-                return null;
-            }
-        }
-
-        private static IEnumerable<Expression> ExtractValueFromNullableArguments(IEnumerable<Expression> arguments)
-        {
-            return arguments.Select(arg => ExtractValueFromNullableExpression(arg));
-        }
-
-        internal static Expression ExtractValueFromNullableExpression(Expression source)
-        {
-            return Nullable.GetUnderlyingType(source.Type) != null ? Expression.Property(source, "Value") : source;
-        }
-
-        internal Expression BindHas(Expression left, Expression flag)
-        {
-            Contract.Assert(TypeHelper.IsEnum(left.Type));
-            Contract.Assert(flag.Type == typeof(Enum));
-
-            Expression[] arguments = new[] { left, flag };
-            return MakeFunctionCall(ClrCanonicalFunctions.HasFlag, arguments);
-        }
-
-        /// <summary>
-        /// Analyze previous query and extract grouped properties.
-        /// </summary>
-        /// <param name="source"></param>
-        protected void EnsureFlattenedPropertyContainer(ParameterExpression source)
-        {
-            if (this.BaseQuery != null)
-            {
-                this.HasInstancePropertyContainer = this.BaseQuery.ElementType.IsGenericType
-                    && this.BaseQuery.ElementType.GetGenericTypeDefinition() == typeof(ComputeWrapper<>);
-
-                this.FlattenedPropertyContainer = this.FlattenedPropertyContainer ?? this.GetFlattenedProperties(source);
-            }
-        }
-
-        internal IDictionary<string, Expression> GetFlattenedProperties(ParameterExpression source)
-        {
-            if (this.BaseQuery == null)
-            {
-                return null;
-            }
-
-            if (!typeof(GroupByWrapper).IsAssignableFrom(BaseQuery.ElementType))
-            {
-                return null;
-            }
-
-            var expression = BaseQuery.Expression as MethodCallExpression;
-            if (expression == null)
-            {
-                return null;
-            }
-
-            // After $apply we could have other clauses, like $filter, $orderby etc.
-            // Skip of filter expressions
-            expression = SkipFilters(expression);
-
-            if (expression == null)
-            {
-                return null;
-            }
-
-            var result = new Dictionary<string, Expression>();
-            CollectContainerAssignments(source, expression, result);
-            if (this.HasInstancePropertyContainer)
-            {
-                var instanceProperty = Expression.Property(source, "Instance");
-                if (typeof(DynamicTypeWrapper).IsAssignableFrom(instanceProperty.Type))
-                {
-                    var computeExpression = expression.Arguments.FirstOrDefault() as MethodCallExpression;
-                    computeExpression = SkipFilters(computeExpression);
-                    if (computeExpression != null)
-                    {
-                        CollectContainerAssignments(instanceProperty, computeExpression, result);
-                    }
-                }
-            }
-
-            return result;
-        }
-
-        private static MethodCallExpression SkipFilters(MethodCallExpression expression)
-        {
-            while (expression.Method.Name == "Where")
-            {
-                expression = expression.Arguments.FirstOrDefault() as MethodCallExpression;
-            }
-
-            return expression;
-        }
-
-        private static void CollectContainerAssignments(Expression source, MethodCallExpression expression, Dictionary<string, Expression> result)
-        {
-            CollectAssigments(result, Expression.Property(source, "GroupByContainer"), ExtractContainerExpression(expression.Arguments.FirstOrDefault() as MethodCallExpression, "GroupByContainer"));
-            CollectAssigments(result, Expression.Property(source, "Container"), ExtractContainerExpression(expression, "Container"));
-        }
-
-        private static MemberInitExpression ExtractContainerExpression(MethodCallExpression expression, string containerName)
-        {
-            if (expression == null || expression.Arguments.Count < 2)
-            {
-                return null;
-            }
-
-            var memberInitExpression = ((expression.Arguments[1] as UnaryExpression).Operand as LambdaExpression).Body as MemberInitExpression;
-            if (memberInitExpression != null)
-            {
-                var containerAssigment = memberInitExpression.Bindings.FirstOrDefault(m => m.Member.Name == containerName) as MemberAssignment;
-                if (containerAssigment != null)
-                {
-                    return containerAssigment.Expression as MemberInitExpression;
-                }
-            }
-            return null;
-        }
-
-        private static void CollectAssigments(IDictionary<string, Expression> flattenPropertyContainer, Expression source, MemberInitExpression expression, string prefix = null)
-        {
-            if (expression == null)
-            {
-                return;
-            }
-
-            string nameToAdd = null;
-            Type resultType = null;
-            MemberInitExpression nextExpression = null;
-            Expression nestedExpression = null;
-            foreach (var expr in expression.Bindings.OfType<MemberAssignment>())
-            {
-                var initExpr = expr.Expression as MemberInitExpression;
-                if (initExpr != null && expr.Member.Name == "Next")
-                {
-                    nextExpression = initExpr;
-                }
-                else if (expr.Member.Name == "Name")
-                {
-                    nameToAdd = (expr.Expression as ConstantExpression).Value as string;
-                }
-                else if (expr.Member.Name == "Value" || expr.Member.Name == "NestedValue")
-                {
-                    resultType = expr.Expression.Type;
-                    if (resultType == typeof(object) && expr.Expression.NodeType == ExpressionType.Convert)
-                    {
-                        resultType = ((UnaryExpression)expr.Expression).Operand.Type;
-                    }
-
-                    if (typeof(GroupByWrapper).IsAssignableFrom(resultType))
-                    {
-                        nestedExpression = expr.Expression;
-                    }
-                }
-            }
-
-            if (prefix != null)
-            {
-                nameToAdd = prefix + "\\" + nameToAdd;
-            }
-
-            if (typeof(GroupByWrapper).IsAssignableFrom(resultType))
-            {
-                flattenPropertyContainer.Add(nameToAdd, Expression.Property(source, "NestedValue"));
-            }
-            else
-            {
-                flattenPropertyContainer.Add(nameToAdd, Expression.Convert(Expression.Property(source, "Value"), resultType));
-            }
-
-            if (nextExpression != null)
-            {
-                CollectAssigments(flattenPropertyContainer, Expression.Property(source, "Next"), nextExpression, prefix);
-            }
-
-            if (nestedExpression != null)
-            {
-                var nestedAccessor = ((nestedExpression as MemberInitExpression).Bindings.First() as MemberAssignment).Expression as MemberInitExpression;
-                var newSource = Expression.Property(Expression.Property(source, "NestedValue"), "GroupByContainer");
-                CollectAssigments(flattenPropertyContainer, newSource, nestedAccessor, nameToAdd);
-            }
-        }
-
-        /// <summary>
-        /// Gets expression for property from previously aggregated query
-        /// </summary>
-        /// <param name="propertyPath"></param>
-        /// <returns>Returns null if no aggregations were used so far</returns>
-        protected Expression GetFlattenedPropertyExpression(string propertyPath)
-        {
-            if (FlattenedPropertyContainer == null)
-            {
-                return null;
-            }
-
-            Expression expression;
-            if (FlattenedPropertyContainer.TryGetValue(propertyPath, out expression))
-            {
-                return expression;
-            }
-
-            if (this.HasInstancePropertyContainer)
-            {
-                return null;
-            }
-
-            throw new ODataException(Error.Format(SRResources.PropertyOrPathWasRemovedFromContext, propertyPath));
-        }
-
-        private Expression GetProperty(Expression source, string propertyName)
-        {
-            if (IsDateOrOffset(source.Type))
-            {
-                if (IsDateTime(source.Type))
-                {
-                    return MakePropertyAccess(ClrCanonicalFunctions.DateTimeProperties[propertyName], source);
-                }
-                else
-                {
-                    return MakePropertyAccess(ClrCanonicalFunctions.DateTimeOffsetProperties[propertyName], source);
-                }
-            }
-            else if (IsDate(source.Type))
-            {
-                return MakePropertyAccess(ClrCanonicalFunctions.DateProperties[propertyName], source);
-            }
-            else if (IsTimeOfDay(source.Type))
-            {
-                return MakePropertyAccess(ClrCanonicalFunctions.TimeOfDayProperties[propertyName], source);
-            }
-            else if (IsTimeSpan(source.Type))
-            {
-                return MakePropertyAccess(ClrCanonicalFunctions.TimeSpanProperties[propertyName], source);
-            }
-
-            return source;
-        }
-
-        private Expression CreateDateBinaryExpression(Expression source)
-        {
-            source = ConvertToDateTimeRelatedConstExpression(source);
-
-            // Year, Month, Day
-            Expression year = GetProperty(source, ClrCanonicalFunctions.YearFunctionName);
-            Expression month = GetProperty(source, ClrCanonicalFunctions.MonthFunctionName);
-            Expression day = GetProperty(source, ClrCanonicalFunctions.DayFunctionName);
-
-            // return (year * 10000 + month * 100 + day)
-            Expression result =
-                Expression.Add(
-                    Expression.Add(Expression.Multiply(year, Expression.Constant(10000)),
-                        Expression.Multiply(month, Expression.Constant(100))), day);
-
-            return CreateFunctionCallWithNullPropagation(result, new[] { source });
-        }
-
-        private Expression CreateTimeBinaryExpression(Expression source)
-        {
-            source = ConvertToDateTimeRelatedConstExpression(source);
-
-            // Hour, Minute, Second, Millisecond
-            Expression hour = GetProperty(source, ClrCanonicalFunctions.HourFunctionName);
-            Expression minute = GetProperty(source, ClrCanonicalFunctions.MinuteFunctionName);
-            Expression second = GetProperty(source, ClrCanonicalFunctions.SecondFunctionName);
-            Expression milliSecond = GetProperty(source, ClrCanonicalFunctions.MillisecondFunctionName);
-
-            Expression hourTicks = Expression.Multiply(Expression.Convert(hour, typeof(long)), Expression.Constant(TimeOfDay.TicksPerHour));
-            Expression minuteTicks = Expression.Multiply(Expression.Convert(minute, typeof(long)), Expression.Constant(TimeOfDay.TicksPerMinute));
-            Expression secondTicks = Expression.Multiply(Expression.Convert(second, typeof(long)), Expression.Constant(TimeOfDay.TicksPerSecond));
-
-            // return (hour * TicksPerHour + minute * TicksPerMinute + second * TicksPerSecond + millisecond)
-            Expression result = Expression.Add(hourTicks, Expression.Add(minuteTicks, Expression.Add(secondTicks, Expression.Convert(milliSecond, typeof(long)))));
-
-            return CreateFunctionCallWithNullPropagation(result, new[] { source });
-        }
-
-        private static Expression ConvertToDateTimeRelatedConstExpression(Expression source)
-        {
-            var parameterizedConstantValue = ExtractParameterizedConstant(source);
-            if (parameterizedConstantValue != null && TypeHelper.IsNullable(source.Type))
-            {
-                var dateTimeOffset = parameterizedConstantValue as DateTimeOffset?;
-                if (dateTimeOffset != null)
-                {
-                    return Expression.Constant(dateTimeOffset.Value, typeof(DateTimeOffset));
-                }
-
-                var dateTime = parameterizedConstantValue as DateTime?;
-                if (dateTime != null)
-                {
-                    return Expression.Constant(dateTime.Value, typeof(DateTime));
-                }
-
-                var date = parameterizedConstantValue as Date?;
-                if (date != null)
-                {
-                    return Expression.Constant(date.Value, typeof(Date));
-                }
-
-                var timeOfDay = parameterizedConstantValue as TimeOfDay?;
-                if (timeOfDay != null)
-                {
-                    return Expression.Constant(timeOfDay.Value, typeof(TimeOfDay));
-                }
-            }
-
-            return source;
-        }
-
-        internal static Expression ConvertToEnumUnderlyingType(Expression expression, Type enumType, Type enumUnderlyingType)
-        {
-            object parameterizedConstantValue = ExtractParameterizedConstant(expression);
-            if (parameterizedConstantValue != null)
-            {
-                string enumStringValue = parameterizedConstantValue as string;
-                if (enumStringValue != null)
-                {
-                    return Expression.Constant(
-                        Convert.ChangeType(
-                            Enum.Parse(enumType, enumStringValue), enumUnderlyingType, CultureInfo.InvariantCulture));
-                }
-                else
-                {
-                    // enum member value
-                    return Expression.Constant(
-                        Convert.ChangeType(
-                            parameterizedConstantValue, enumUnderlyingType, CultureInfo.InvariantCulture));
-                }
-            }
-            else if (expression.Type == enumType)
-            {
-                return Expression.Convert(expression, enumUnderlyingType);
-            }
-            else if (Nullable.GetUnderlyingType(expression.Type) == enumType)
-            {
-                return Expression.Convert(expression, typeof(Nullable<>).MakeGenericType(enumUnderlyingType));
-            }
-            else if (expression.NodeType == ExpressionType.Constant && ((ConstantExpression)expression).Value == null)
-            {
-                return expression;
-            }
-            else
-            {
-                throw Error.NotSupported(SRResources.ConvertToEnumFailed, enumType, expression.Type);
-            }
-        }
-
-        // Extract the constant that would have been encapsulated into LinqParameterContainer if this
-        // expression represents it else return null.
-        internal static object ExtractParameterizedConstant(Expression expression)
-        {
-            if (expression.NodeType == ExpressionType.MemberAccess)
-            {
-                MemberExpression memberAccess = expression as MemberExpression;
-                Contract.Assert(memberAccess != null);
-
-                PropertyInfo propertyInfo = memberAccess.Member as PropertyInfo;
-                if (propertyInfo != null && propertyInfo.GetMethod.IsStatic)
-                {
-                    return propertyInfo.GetValue(new object());
-                }
-
-                if (memberAccess.Expression.NodeType == ExpressionType.Constant)
-                {
-                    ConstantExpression constant = memberAccess.Expression as ConstantExpression;
-                    Contract.Assert(constant != null);
-                    Contract.Assert(constant.Value != null);
-                    LinqParameterContainer value = constant.Value as LinqParameterContainer;
-                    Contract.Assert(value != null, "Constants are already embedded into LinqParameterContainer");
-
-                    return value.Property;
-                }
-            }
-
-            return null;
-        }
-
-        internal static Expression DateTimeOffsetToDateTime(Expression expression, TimeZoneInfo timeZoneInfo)
-        {
-            var unaryExpression = expression as UnaryExpression;
-            if (unaryExpression != null)
-            {
-                if (Nullable.GetUnderlyingType(unaryExpression.Type) == unaryExpression.Operand.Type)
-                {
-                    // this is a cast from T to Nullable<T> which is redundant.
-                    expression = unaryExpression.Operand;
-                }
-            }
-            var parameterizedConstantValue = ExtractParameterizedConstant(expression);
-            var dto = parameterizedConstantValue as DateTimeOffset?;
-            if (dto != null)
-            {
-                expression = Expression.Constant(EdmPrimitiveHelper.ConvertPrimitiveValue(dto.Value, typeof(DateTime), timeZoneInfo));
-            }
-            return expression;
-        }
-
-        internal static bool IsNullable(Type t)
-        {
-            if (!t.IsValueType || (t.IsGenericType && t.GetGenericTypeDefinition() == typeof(Nullable<>)))
-            {
-                return true;
-            }
-
-            return false;
-        }
-
-        internal static Type ToNullable(Type t)
-        {
-            if (IsNullable(t))
-            {
-                return t;
-            }
-            else
-            {
-                return typeof(Nullable<>).MakeGenericType(t);
-            }
-        }
-
-        internal static Expression ToNullable(Expression expression)
-        {
-            if (!IsNullable(expression.Type))
-            {
-                return Expression.Convert(expression, ToNullable(expression.Type));
-            }
-
-            return expression;
-        }
-
-        internal static bool IsIQueryable(Type type)
-        {
-            return typeof(IQueryable).IsAssignableFrom(type);
-        }
-
-        internal static bool IsDoubleOrDecimal(Type type)
-        {
-            return IsType<double>(type) || IsType<decimal>(type);
-        }
-
-        internal static bool IsDateAndTimeRelated(Type type)
-        {
-            return IsType<Date>(type) ||
-                IsType<DateTime>(type) ||
-                IsType<DateTimeOffset>(type) ||
-                IsType<TimeOfDay>(type) ||
-                IsType<TimeSpan>(type);
-        }
-
-        internal static bool IsDateRelated(Type type)
-        {
-            return IsType<Date>(type) || IsType<DateTime>(type) || IsType<DateTimeOffset>(type);
-        }
-
-        internal static bool IsTimeRelated(Type type)
-        {
-            return IsType<TimeOfDay>(type) || IsType<DateTime>(type) || IsType<DateTimeOffset>(type) || IsType<TimeSpan>(type);
-        }
-
-        internal static bool IsDateOrOffset(Type type)
-        {
-            return IsType<DateTime>(type) || IsType<DateTimeOffset>(type);
-        }
-
-        internal static bool IsDateTime(Type type)
-        {
-            return IsType<DateTime>(type);
-        }
-
-        internal static bool IsTimeSpan(Type type)
-        {
-            return IsType<TimeSpan>(type);
-        }
-
-        internal static bool IsTimeOfDay(Type type)
-        {
-            return IsType<TimeOfDay>(type);
-        }
-
-        internal static bool IsDate(Type type)
-        {
-            return IsType<Date>(type);
-        }
-
-        internal static bool IsInteger(Type type)
-        {
-            return IsType<short>(type) || IsType<int>(type) || IsType<long>(type);
-        }
-
-        internal static bool IsType<T>(Type type) where T : struct
-        {
-            return type == typeof(T) || type == typeof(T?);
-        }
-
-        internal static Expression ConvertNull(Expression expression, Type type)
-        {
-            ConstantExpression constantExpression = expression as ConstantExpression;
-            if (constantExpression != null && constantExpression.Value == null)
-            {
-                return Expression.Constant(null, type);
-            }
-            else
-            {
-                return expression;
-            }
-        }
-
-        /// <summary>
-        /// Recognize $it.Source where $it is FlatteningWrapper
-        /// Using that do avoid wrapping it redundant into Null propagation 
-        /// </summary>
-        /// <param name="source"></param>
-        /// <returns></returns>
-        private bool IsFlatteningSource(Expression source)
-        {
-            var member = source as MemberExpression;
-            return member != null
-                && this.Parameter.Type.IsGenericType
-                && this.Parameter.Type.GetGenericTypeDefinition() == typeof(FlatteningWrapper<>)
-                && member.Expression == this.Parameter;
-        }
-
-        internal Expression CreatePropertyAccessExpression(Expression source, IEdmProperty property, string propertyPath = null)
-        {
-            string propertyName = Model.GetClrPropertyName(property);
-            propertyPath = propertyPath ?? propertyName;
-            if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True && IsNullable(source.Type) &&
-                source != this.Parameter &&
-                !IsFlatteningSource(source))
-            {
-                Expression cleanSource = RemoveInnerNullPropagation(source);
-                Expression propertyAccessExpression = null;
-                propertyAccessExpression = GetFlattenedPropertyExpression(propertyPath) ?? Expression.Property(cleanSource, propertyName);
-
-                // source.property => source == null ? null : [CastToNullable]RemoveInnerNullPropagation(source).property
-                // Notice that we are checking if source is null already. so we can safely remove any null checks when doing source.Property
-
-                Expression ifFalse = ToNullable(ConvertNonStandardPrimitives(propertyAccessExpression));
-                return
-                    Expression.Condition(
-                        test: Expression.Equal(source, NullConstant),
-                        ifTrue: Expression.Constant(null, ifFalse.Type),
-                        ifFalse: ifFalse);
-            }
-            else
-            {
-                return GetFlattenedPropertyExpression(propertyPath)
-                    ?? ConvertNonStandardPrimitives(GetPropertyExpression(source, (this.HasInstancePropertyContainer && !propertyPath.Contains("\\", StringComparison.Ordinal) ? "Instance\\" : String.Empty) + propertyName));
-            }
-        }
-
-        internal static Expression GetPropertyExpression(Expression source, string propertyPath)
-        {
-            string[] propertyNameParts = propertyPath.Split('\\');
-            Expression propertyValue = source;
-            foreach (var propertyName in propertyNameParts)
-            {
-                propertyValue = Expression.Property(propertyValue, propertyName);
-            }
-            return propertyValue;
-        }
-
         /// <summary>
         /// Binds a <see cref="QueryNode"/> to create a LINQ <see cref="Expression"/> that represents the semantics
         /// of the <see cref="QueryNode"/>.
@@ -1367,145 +319,6 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
             return Expression.Constant(castedList, listType);
         }
 
-        private Expression BindCastSingleValue(SingleValueFunctionCallNode node)
-        {
-            Contract.Assert(ClrCanonicalFunctions.CastFunctionName == node.Name);
-
-            Expression[] arguments = BindArguments(node.Parameters);
-            Contract.Assert(arguments.Length == 1 || arguments.Length == 2);
-
-            Expression source = arguments.Length == 1 ? this.Parameter : arguments[0];
-            string targetTypeName = (string)((ConstantNode)node.Parameters.Last()).Value;
-            IEdmType targetEdmType = Model.FindType(targetTypeName);
-            Type targetClrType = null;
-
-            if (targetEdmType != null)
-            {
-                IEdmTypeReference targetEdmTypeReference = targetEdmType.ToEdmTypeReference(false);
-                targetClrType = Model.GetClrType(targetEdmTypeReference);
-
-                if (source != NullConstant)
-                {
-                    if (source.Type == targetClrType)
-                    {
-                        return source;
-                    }
-
-                    if ((!targetEdmTypeReference.IsPrimitive() && !targetEdmTypeReference.IsEnum()) ||
-                        (source.Type.GetEdmPrimitiveType() == null && !TypeHelper.IsEnum(source.Type)))
-                    {
-                        // Cast fails and return null.
-                        return NullConstant;
-                    }
-                }
-            }
-
-            if (targetClrType == null || source == NullConstant)
-            {
-                return NullConstant;
-            }
-
-            if (targetClrType == typeof(string))
-            {
-                return BindCastToStringType(source);
-            }
-            else if (TypeHelper.IsEnum(targetClrType))
-            {
-                return BindCastToEnumType(source.Type, targetClrType, node.Parameters.First(), arguments.Length);
-            }
-            else
-            {
-                if (TypeHelper.IsNullable(source.Type) && !TypeHelper.IsNullable(targetClrType))
-                {
-                    // Make the target Clr type nullable to avoid failure while casting
-                    // nullable source, whose value may be null, to a non-nullable type.
-                    // For example: cast(NullableInt32Property,Edm.Int64)
-                    // The target Clr type should be Nullable<Int64> rather than Int64.
-                    targetClrType = typeof(Nullable<>).MakeGenericType(targetClrType);
-                }
-
-                try
-                {
-                    return Expression.Convert(source, targetClrType);
-                }
-                catch (InvalidOperationException)
-                {
-                    // Cast fails and return null.
-                    return NullConstant;
-                }
-            }
-        }
-
-        private static Expression BindCastToStringType(Expression source)
-        {
-            Expression sourceValue;
-
-            if (source.Type.IsGenericType && source.Type.GetGenericTypeDefinition() == typeof(Nullable<>))
-            {
-                if (TypeHelper.IsEnum(source.Type))
-                {
-                    // Entity Framework doesn't have ToString method for enum types.
-                    // Convert enum types to their underlying numeric types.
-                    sourceValue = Expression.Convert(
-                        Expression.Property(source, "Value"),
-                        Enum.GetUnderlyingType(TypeHelper.GetUnderlyingTypeOrSelf(source.Type)));
-                }
-                else
-                {
-                    // Entity Framework has ToString method for numeric types.
-                    sourceValue = Expression.Property(source, "Value");
-                }
-
-                // Entity Framework doesn't have ToString method for nullable numeric types.
-                // Call ToString method on non-nullable numeric types.
-                return Expression.Condition(
-                    Expression.Property(source, "HasValue"),
-                    Expression.Call(sourceValue, "ToString", typeArguments: null, arguments: null),
-                    Expression.Constant(null, typeof(string)));
-            }
-            else
-            {
-                sourceValue = TypeHelper.IsEnum(source.Type) ?
-                    Expression.Convert(source, Enum.GetUnderlyingType(source.Type)) :
-                    source;
-                return Expression.Call(sourceValue, "ToString", typeArguments: null, arguments: null);
-            }
-        }
-
-        private Expression BindCastToEnumType(Type sourceType, Type targetClrType, QueryNode firstParameter, int parameterLength)
-        {
-            Type enumType = TypeHelper.GetUnderlyingTypeOrSelf(targetClrType);
-            ConstantNode sourceNode = firstParameter as ConstantNode;
-
-            if (parameterLength == 1 || sourceNode == null || sourceType != typeof(string))
-            {
-                // We only support to cast Enumeration type from constant string now,
-                // because LINQ to Entities does not recognize the method Enum.TryParse.
-                return NullConstant;
-            }
-            else
-            {
-                object[] parameters = new[] { sourceNode.Value, Enum.ToObject(enumType, 0) };
-                bool isSuccessful = (bool)EnumTryParseMethod.MakeGenericMethod(enumType).Invoke(null, parameters);
-
-                if (isSuccessful)
-                {
-                    if (QuerySettings.EnableConstantParameterization)
-                    {
-                        return LinqParameterContainer.Parameterize(targetClrType, parameters[1]);
-                    }
-                    else
-                    {
-                        return Expression.Constant(parameters[1], targetClrType);
-                    }
-                }
-                else
-                {
-                    return NullConstant;
-                }
-            }
-        }
-
         private Expression BindIsOf(SingleValueFunctionCallNode node)
         {
             Contract.Assert(ClrCanonicalFunctions.IsofFunctionName == node.Name);
@@ -1562,12 +375,12 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
 
             Expression[] arguments = BindArguments(node.Parameters);
 
-            Contract.Assert(arguments.Length == 1 && IsDoubleOrDecimal(arguments[0].Type));
+            Contract.Assert(arguments.Length == 1 && ExpressionBinderHelper.IsDoubleOrDecimal(arguments[0].Type));
 
-            MethodInfo ceiling = IsType<double>(arguments[0].Type)
+            MethodInfo ceiling = ExpressionBinderHelper.IsType<double>(arguments[0].Type)
                 ? ClrCanonicalFunctions.CeilingOfDouble
                 : ClrCanonicalFunctions.CeilingOfDecimal;
-            return MakeFunctionCall(ceiling, arguments);
+            return ExpressionBinderHelper.MakeFunctionCall(ceiling, QuerySettings, arguments);
         }
 
         private Expression BindFloor(SingleValueFunctionCallNode node)
@@ -1576,12 +389,12 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
 
             Expression[] arguments = BindArguments(node.Parameters);
 
-            Contract.Assert(arguments.Length == 1 && IsDoubleOrDecimal(arguments[0].Type));
+            Contract.Assert(arguments.Length == 1 && ExpressionBinderHelper.IsDoubleOrDecimal(arguments[0].Type));
 
-            MethodInfo floor = IsType<double>(arguments[0].Type)
+            MethodInfo floor = ExpressionBinderHelper.IsType<double>(arguments[0].Type)
                 ? ClrCanonicalFunctions.FloorOfDouble
                 : ClrCanonicalFunctions.FloorOfDecimal;
-            return MakeFunctionCall(floor, arguments);
+            return ExpressionBinderHelper.MakeFunctionCall(floor, QuerySettings, arguments);
         }
 
         private Expression BindRound(SingleValueFunctionCallNode node)
@@ -1590,12 +403,12 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
 
             Expression[] arguments = BindArguments(node.Parameters);
 
-            Contract.Assert(arguments.Length == 1 && IsDoubleOrDecimal(arguments[0].Type));
+            Contract.Assert(arguments.Length == 1 && ExpressionBinderHelper.IsDoubleOrDecimal(arguments[0].Type));
 
-            MethodInfo round = IsType<double>(arguments[0].Type)
+            MethodInfo round = ExpressionBinderHelper.IsType<double>(arguments[0].Type)
                 ? ClrCanonicalFunctions.RoundOfDouble
                 : ClrCanonicalFunctions.RoundOfDecimal;
-            return MakeFunctionCall(round, arguments);
+            return ExpressionBinderHelper.MakeFunctionCall(round, QuerySettings, arguments);
         }
 
         private Expression BindDate(SingleValueFunctionCallNode node)
@@ -1605,7 +418,7 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
             Expression[] arguments = BindArguments(node.Parameters);
 
             // We should support DateTime & DateTimeOffset even though DateTime is not part of OData v4 Spec.
-            Contract.Assert(arguments.Length == 1 && IsDateOrOffset(arguments[0].Type));
+            Contract.Assert(arguments.Length == 1 && ExpressionBinderHelper.IsDateOrOffset(arguments[0].Type));
 
             // EF doesn't support new Date(int, int, int), also doesn't support other property access, for example DateTime.Date.
             // Therefore, we just return the source (DateTime or DateTimeOffset).
@@ -1630,7 +443,7 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
             Expression[] arguments = BindArguments(node.Parameters);
 
             // We should support DateTime & DateTimeOffset even though DateTime is not part of OData v4 Spec.
-            Contract.Assert(arguments.Length == 1 && IsDateOrOffset(arguments[0].Type));
+            Contract.Assert(arguments.Length == 1 && ExpressionBinderHelper.IsDateOrOffset(arguments[0].Type));
 
             // EF doesn't support new TimeOfDay(int, int, int, int), also doesn't support other property access, for example DateTimeOffset.DateTime.
             // Therefore, we just return the source (DateTime or DateTimeOffset).
@@ -1642,21 +455,21 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
             Contract.Assert("fractionalseconds" == node.Name);
 
             Expression[] arguments = BindArguments(node.Parameters);
-            Contract.Assert(arguments.Length == 1 && (IsTimeRelated(arguments[0].Type)));
+            Contract.Assert(arguments.Length == 1 && (ExpressionBinderHelper.IsTimeRelated(arguments[0].Type)));
 
             // We should support DateTime & DateTimeOffset even though DateTime is not part of OData v4 Spec.
             Expression parameter = arguments[0];
 
             PropertyInfo property;
-            if (IsTimeOfDay(parameter.Type))
+            if (ExpressionBinderHelper.IsTimeOfDay(parameter.Type))
             {
                 property = ClrCanonicalFunctions.TimeOfDayProperties[ClrCanonicalFunctions.MillisecondFunctionName];
             }
-            else if (IsDateTime(parameter.Type))
+            else if (ExpressionBinderHelper.IsDateTime(parameter.Type))
             {
                 property = ClrCanonicalFunctions.DateTimeProperties[ClrCanonicalFunctions.MillisecondFunctionName];
             }
-            else if (IsTimeSpan(parameter.Type))
+            else if (ExpressionBinderHelper.IsTimeSpan(parameter.Type))
             {
                 property = ClrCanonicalFunctions.TimeSpanProperties[ClrCanonicalFunctions.MillisecondFunctionName];
             }
@@ -1666,28 +479,28 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
             }
 
             // Millisecond
-            Expression milliSecond = MakePropertyAccess(property, parameter);
+            Expression milliSecond = ExpressionBinderHelper.MakePropertyAccess(property, parameter, QuerySettings);
             Expression decimalMilliSecond = Expression.Convert(milliSecond, typeof(decimal));
             Expression fractionalSeconds = Expression.Divide(decimalMilliSecond, Expression.Constant(1000m, typeof(decimal)));
 
-            return CreateFunctionCallWithNullPropagation(fractionalSeconds, arguments);
+            return ExpressionBinderHelper.CreateFunctionCallWithNullPropagation(fractionalSeconds, arguments, QuerySettings);
         }
 
         private Expression BindDateRelatedProperty(SingleValueFunctionCallNode node)
         {
             Expression[] arguments = BindArguments(node.Parameters);
-            Contract.Assert(arguments.Length == 1 && IsDateRelated(arguments[0].Type));
+            Contract.Assert(arguments.Length == 1 && ExpressionBinderHelper.IsDateRelated(arguments[0].Type));
 
             // We should support DateTime & DateTimeOffset even though DateTime is not part of OData v4 Spec.
             Expression parameter = arguments[0];
 
             PropertyInfo property;
-            if (IsDate(parameter.Type))
+            if (ExpressionBinderHelper.IsDate(parameter.Type))
             {
                 Contract.Assert(ClrCanonicalFunctions.DateProperties.ContainsKey(node.Name));
                 property = ClrCanonicalFunctions.DateProperties[node.Name];
             }
-            else if (IsDateTime(parameter.Type))
+            else if (ExpressionBinderHelper.IsDateTime(parameter.Type))
             {
                 Contract.Assert(ClrCanonicalFunctions.DateTimeProperties.ContainsKey(node.Name));
                 property = ClrCanonicalFunctions.DateTimeProperties[node.Name];
@@ -1698,29 +511,29 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
                 property = ClrCanonicalFunctions.DateTimeOffsetProperties[node.Name];
             }
 
-            return MakeFunctionCall(property, parameter);
+            return ExpressionBinderHelper.MakeFunctionCall(property, QuerySettings, parameter);
         }
 
         private Expression BindTimeRelatedProperty(SingleValueFunctionCallNode node)
         {
             Expression[] arguments = BindArguments(node.Parameters);
-            Contract.Assert(arguments.Length == 1 && (IsTimeRelated(arguments[0].Type)));
+            Contract.Assert(arguments.Length == 1 && (ExpressionBinderHelper.IsTimeRelated(arguments[0].Type)));
 
             // We should support DateTime & DateTimeOffset even though DateTime is not part of OData v4 Spec.
             Expression parameter = arguments[0];
 
             PropertyInfo property;
-            if (IsTimeOfDay(parameter.Type))
+            if (ExpressionBinderHelper.IsTimeOfDay(parameter.Type))
             {
                 Contract.Assert(ClrCanonicalFunctions.TimeOfDayProperties.ContainsKey(node.Name));
                 property = ClrCanonicalFunctions.TimeOfDayProperties[node.Name];
             }
-            else if (IsDateTime(parameter.Type))
+            else if (ExpressionBinderHelper.IsDateTime(parameter.Type))
             {
                 Contract.Assert(ClrCanonicalFunctions.DateTimeProperties.ContainsKey(node.Name));
                 property = ClrCanonicalFunctions.DateTimeProperties[node.Name];
             }
-            else if (IsTimeSpan(parameter.Type))
+            else if (ExpressionBinderHelper.IsTimeSpan(parameter.Type))
             {
                 Contract.Assert(ClrCanonicalFunctions.TimeSpanProperties.ContainsKey(node.Name));
                 property = ClrCanonicalFunctions.TimeSpanProperties[node.Name];
@@ -1731,7 +544,7 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
                 property = ClrCanonicalFunctions.DateTimeOffsetProperties[node.Name];
             }
 
-            return MakeFunctionCall(property, parameter);
+            return ExpressionBinderHelper.MakeFunctionCall(property, QuerySettings, parameter);
         }
 
         private Expression BindConcat(SingleValueFunctionCallNode node)
@@ -1743,7 +556,7 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
 
             Contract.Assert(arguments.Length == 2 && arguments[0].Type == typeof(string) && arguments[1].Type == typeof(string));
 
-            return MakeFunctionCall(ClrCanonicalFunctions.Concat, arguments);
+            return ExpressionBinderHelper.MakeFunctionCall(ClrCanonicalFunctions.Concat, QuerySettings, arguments);
         }
 
         private Expression BindTrim(SingleValueFunctionCallNode node)
@@ -1755,7 +568,7 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
 
             Contract.Assert(arguments.Length == 1 && arguments[0].Type == typeof(string));
 
-            return MakeFunctionCall(ClrCanonicalFunctions.Trim, arguments);
+            return ExpressionBinderHelper.MakeFunctionCall(ClrCanonicalFunctions.Trim, QuerySettings, arguments);
         }
 
         private Expression BindToUpper(SingleValueFunctionCallNode node)
@@ -1767,7 +580,7 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
 
             Contract.Assert(arguments.Length == 1 && arguments[0].Type == typeof(string));
 
-            return MakeFunctionCall(ClrCanonicalFunctions.ToUpper, arguments);
+            return ExpressionBinderHelper.MakeFunctionCall(ClrCanonicalFunctions.ToUpper, QuerySettings, arguments);
         }
 
         private Expression BindToLower(SingleValueFunctionCallNode node)
@@ -1779,7 +592,7 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
 
             Contract.Assert(arguments.Length == 1 && arguments[0].Type == typeof(string));
 
-            return MakeFunctionCall(ClrCanonicalFunctions.ToLower, arguments);
+            return ExpressionBinderHelper.MakeFunctionCall(ClrCanonicalFunctions.ToLower, QuerySettings, arguments);
         }
 
         private Expression BindIndexOf(SingleValueFunctionCallNode node)
@@ -1791,7 +604,7 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
 
             Contract.Assert(arguments.Length == 2 && arguments[0].Type == typeof(string) && arguments[1].Type == typeof(string));
 
-            return MakeFunctionCall(ClrCanonicalFunctions.IndexOf, arguments);
+            return ExpressionBinderHelper.MakeFunctionCall(ClrCanonicalFunctions.IndexOf, QuerySettings, arguments);
         }
 
         private Expression BindSubstring(SingleValueFunctionCallNode node)
@@ -1807,7 +620,7 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
             Expression functionCall;
             if (arguments.Length == 2)
             {
-                Contract.Assert(IsInteger(arguments[1].Type));
+                Contract.Assert(ExpressionBinderHelper.IsInteger(arguments[1].Type));
 
                 // When null propagation is allowed, we use a safe version of String.Substring(int).
                 // But for providers that would not recognize custom expressions like this, we map
@@ -1815,17 +628,17 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
                 if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True)
                 {
                     // Safe function is static and takes string "this" as first argument
-                    functionCall = MakeFunctionCall(ClrCanonicalFunctions.SubstringStartNoThrow, arguments);
+                    functionCall = ExpressionBinderHelper.MakeFunctionCall(ClrCanonicalFunctions.SubstringStartNoThrow, QuerySettings, arguments);
                 }
                 else
                 {
-                    functionCall = MakeFunctionCall(ClrCanonicalFunctions.SubstringStart, arguments);
+                    functionCall = ExpressionBinderHelper.MakeFunctionCall(ClrCanonicalFunctions.SubstringStart, QuerySettings, arguments);
                 }
             }
             else
             {
                 // arguments.Length == 3 implies String.Substring(int, int)
-                Contract.Assert(arguments.Length == 3 && IsInteger(arguments[1].Type) && IsInteger(arguments[2].Type));
+                Contract.Assert(arguments.Length == 3 && ExpressionBinderHelper.IsInteger(arguments[1].Type) && ExpressionBinderHelper.IsInteger(arguments[2].Type));
 
                 // When null propagation is allowed, we use a safe version of String.Substring(int, int).
                 // But for providers that would not recognize custom expressions like this, we map
@@ -1833,11 +646,11 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
                 if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True)
                 {
                     // Safe function is static and takes string "this" as first argument
-                    functionCall = MakeFunctionCall(ClrCanonicalFunctions.SubstringStartAndLengthNoThrow, arguments);
+                    functionCall = ExpressionBinderHelper.MakeFunctionCall(ClrCanonicalFunctions.SubstringStartAndLengthNoThrow, QuerySettings, arguments);
                 }
                 else
                 {
-                    functionCall = MakeFunctionCall(ClrCanonicalFunctions.SubstringStartAndLength, arguments);
+                    functionCall = ExpressionBinderHelper.MakeFunctionCall(ClrCanonicalFunctions.SubstringStartAndLength, QuerySettings, arguments);
                 }
             }
 
@@ -1853,7 +666,7 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
 
             Contract.Assert(arguments.Length == 1 && arguments[0].Type == typeof(string));
 
-            return MakeFunctionCall(ClrCanonicalFunctions.Length, arguments);
+            return ExpressionBinderHelper.MakeFunctionCall(ClrCanonicalFunctions.Length, QuerySettings, arguments);
         }
 
         private Expression BindContains(SingleValueFunctionCallNode node)
@@ -1865,7 +678,7 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
 
             Contract.Assert(arguments.Length == 2 && arguments[0].Type == typeof(string) && arguments[1].Type == typeof(string));
 
-            return MakeFunctionCall(ClrCanonicalFunctions.Contains, arguments[0], arguments[1]);
+            return ExpressionBinderHelper.MakeFunctionCall(ClrCanonicalFunctions.Contains, QuerySettings, arguments[0], arguments[1]);
         }
 
         private Expression BindStartsWith(SingleValueFunctionCallNode node)
@@ -1877,7 +690,7 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
 
             Contract.Assert(arguments.Length == 2 && arguments[0].Type == typeof(string) && arguments[1].Type == typeof(string));
 
-            return MakeFunctionCall(ClrCanonicalFunctions.StartsWith, arguments);
+            return ExpressionBinderHelper.MakeFunctionCall(ClrCanonicalFunctions.StartsWith, QuerySettings, arguments);
         }
 
         private Expression BindEndsWith(SingleValueFunctionCallNode node)
@@ -1889,7 +702,213 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
 
             Contract.Assert(arguments.Length == 2 && arguments[0].Type == typeof(string) && arguments[1].Type == typeof(string));
 
-            return MakeFunctionCall(ClrCanonicalFunctions.EndsWith, arguments);
+            return ExpressionBinderHelper.MakeFunctionCall(ClrCanonicalFunctions.EndsWith, QuerySettings, arguments);
+        }
+
+        private Expression BindCustomMethodExpressionOrNull(SingleValueFunctionCallNode node)
+        {
+            Expression[] arguments = BindArguments(node.Parameters);
+            IEnumerable<Type> methodArgumentsType = arguments.Select(argument => argument.Type);
+
+            // Search for custom method info that are binded to the node name
+            MethodInfo methodInfo;
+            if (UriFunctionsBinder.TryGetMethodInfo(node.Name, methodArgumentsType, out methodInfo))
+            {
+                return ExpressionBinderHelper.MakeFunctionCall(methodInfo, QuerySettings, arguments);
+            }
+
+            return null;
+        }
+
+        private Expression BindCastSingleValue(SingleValueFunctionCallNode node)
+        {
+            Contract.Assert(ClrCanonicalFunctions.CastFunctionName == node.Name);
+
+            Expression[] arguments = BindArguments(node.Parameters);
+            Contract.Assert(arguments.Length == 1 || arguments.Length == 2);
+
+            Expression source = arguments.Length == 1 ? this.Parameter : arguments[0];
+            string targetTypeName = (string)((ConstantNode)node.Parameters.Last()).Value;
+            IEdmType targetEdmType = Model.FindType(targetTypeName);
+            Type targetClrType = null;
+
+            if (targetEdmType != null)
+            {
+                IEdmTypeReference targetEdmTypeReference = targetEdmType.ToEdmTypeReference(false);
+                targetClrType = Model.GetClrType(targetEdmTypeReference);
+
+                if (source != NullConstant)
+                {
+                    if (source.Type == targetClrType)
+                    {
+                        return source;
+                    }
+
+                    if ((!targetEdmTypeReference.IsPrimitive() && !targetEdmTypeReference.IsEnum()) ||
+                        (source.Type.GetEdmPrimitiveType() == null && !TypeHelper.IsEnum(source.Type)))
+                    {
+                        // Cast fails and return null.
+                        return NullConstant;
+                    }
+                }
+            }
+
+            if (targetClrType == null || source == NullConstant)
+            {
+                return NullConstant;
+            }
+
+            if (targetClrType == typeof(string))
+            {
+                return ExpressionBinderHelper.BindCastToStringType(source);
+            }
+            else if (TypeHelper.IsEnum(targetClrType))
+            {
+                return BindCastToEnumType(source.Type, targetClrType, node.Parameters.First(), arguments.Length);
+            }
+            else
+            {
+                if (TypeHelper.IsNullable(source.Type) && !TypeHelper.IsNullable(targetClrType))
+                {
+                    // Make the target Clr type nullable to avoid failure while casting
+                    // nullable source, whose value may be null, to a non-nullable type.
+                    // For example: cast(NullableInt32Property,Edm.Int64)
+                    // The target Clr type should be Nullable<Int64> rather than Int64.
+                    targetClrType = typeof(Nullable<>).MakeGenericType(targetClrType);
+                }
+
+                try
+                {
+                    return Expression.Convert(source, targetClrType);
+                }
+                catch (InvalidOperationException)
+                {
+                    // Cast fails and return null.
+                    return NullConstant;
+                }
+            }
+        }
+
+        private static void ValidateAllStringArguments(string functionName, Expression[] arguments)
+        {
+            if (arguments.Any(arg => arg.Type != typeof(string)))
+            {
+                throw new ODataException(Error.Format(SRResources.FunctionNotSupportedOnEnum, functionName));
+            }
+        }
+
+        /// <summary>
+        /// Recognize $it.Source where $it is FlatteningWrapper
+        /// Using that do avoid wrapping it redundant into Null propagation 
+        /// </summary>
+        /// <param name="source"></param>
+        /// <returns></returns>
+        private bool IsFlatteningSource(Expression source)
+        {
+            var member = source as MemberExpression;
+            return member != null
+                && this.Parameter.Type.IsGenericType
+                && this.Parameter.Type.GetGenericTypeDefinition() == typeof(FlatteningWrapper<>)
+                && member.Expression == this.Parameter;
+        }
+
+        private static MethodCallExpression SkipFilters(MethodCallExpression expression)
+        {
+            while (expression.Method.Name == "Where")
+            {
+                expression = expression.Arguments.FirstOrDefault() as MethodCallExpression;
+            }
+
+            return expression;
+        }
+
+        private static void CollectContainerAssignments(Expression source, MethodCallExpression expression, Dictionary<string, Expression> result)
+        {
+            CollectAssigments(result, Expression.Property(source, "GroupByContainer"), ExtractContainerExpression(expression.Arguments.FirstOrDefault() as MethodCallExpression, "GroupByContainer"));
+            CollectAssigments(result, Expression.Property(source, "Container"), ExtractContainerExpression(expression, "Container"));
+        }
+
+        private static void CollectAssigments(IDictionary<string, Expression> flattenPropertyContainer, Expression source, MemberInitExpression expression, string prefix = null)
+        {
+            if (expression == null)
+            {
+                return;
+            }
+
+            string nameToAdd = null;
+            Type resultType = null;
+            MemberInitExpression nextExpression = null;
+            Expression nestedExpression = null;
+            foreach (var expr in expression.Bindings.OfType<MemberAssignment>())
+            {
+                var initExpr = expr.Expression as MemberInitExpression;
+                if (initExpr != null && expr.Member.Name == "Next")
+                {
+                    nextExpression = initExpr;
+                }
+                else if (expr.Member.Name == "Name")
+                {
+                    nameToAdd = (expr.Expression as ConstantExpression).Value as string;
+                }
+                else if (expr.Member.Name == "Value" || expr.Member.Name == "NestedValue")
+                {
+                    resultType = expr.Expression.Type;
+                    if (resultType == typeof(object) && expr.Expression.NodeType == ExpressionType.Convert)
+                    {
+                        resultType = ((UnaryExpression)expr.Expression).Operand.Type;
+                    }
+
+                    if (typeof(GroupByWrapper).IsAssignableFrom(resultType))
+                    {
+                        nestedExpression = expr.Expression;
+                    }
+                }
+            }
+
+            if (prefix != null)
+            {
+                nameToAdd = prefix + "\\" + nameToAdd;
+            }
+
+            if (typeof(GroupByWrapper).IsAssignableFrom(resultType))
+            {
+                flattenPropertyContainer.Add(nameToAdd, Expression.Property(source, "NestedValue"));
+            }
+            else
+            {
+                flattenPropertyContainer.Add(nameToAdd, Expression.Convert(Expression.Property(source, "Value"), resultType));
+            }
+
+            if (nextExpression != null)
+            {
+                CollectAssigments(flattenPropertyContainer, Expression.Property(source, "Next"), nextExpression, prefix);
+            }
+
+            if (nestedExpression != null)
+            {
+                var nestedAccessor = ((nestedExpression as MemberInitExpression).Bindings.First() as MemberAssignment).Expression as MemberInitExpression;
+                var newSource = Expression.Property(Expression.Property(source, "NestedValue"), "GroupByContainer");
+                CollectAssigments(flattenPropertyContainer, newSource, nestedAccessor, nameToAdd);
+            }
+        }
+
+        private static MemberInitExpression ExtractContainerExpression(MethodCallExpression expression, string containerName)
+        {
+            if (expression == null || expression.Arguments.Count < 2)
+            {
+                return null;
+            }
+
+            var memberInitExpression = ((expression.Arguments[1] as UnaryExpression).Operand as LambdaExpression).Body as MemberInitExpression;
+            if (memberInitExpression != null)
+            {
+                var containerAssigment = memberInitExpression.Bindings.FirstOrDefault(m => m.Member.Name == containerName) as MemberAssignment;
+                if (containerAssigment != null)
+                {
+                    return containerAssigment.Expression as MemberInitExpression;
+                }
+            }
+            return null;
         }
 
         /// <summary>
@@ -1902,30 +921,338 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
             return nodes.OfType<SingleValueNode>().Select(n => Bind(n)).ToArray();
         }
 
-        private static void ValidateAllStringArguments(string functionName, Expression[] arguments)
+        /// <summary>
+        /// Gets property for dynamic properties dictionary.
+        /// </summary>
+        /// <param name="openNode"></param>
+        /// <returns>Returns CLR property for dynamic properties container.</returns>
+        protected PropertyInfo GetDynamicPropertyContainer(SingleValueOpenPropertyAccessNode openNode)
         {
-            if (arguments.Any(arg => arg.Type != typeof(string)))
+            if (openNode == null)
             {
-                throw new ODataException(Error.Format(SRResources.FunctionNotSupportedOnEnum, functionName));
+                throw Error.ArgumentNull(nameof(openNode));
+            }
+
+            IEdmStructuredType edmStructuredType;
+            IEdmTypeReference edmTypeReference = openNode.Source.TypeReference;
+            if (edmTypeReference.IsEntity())
+            {
+                edmStructuredType = edmTypeReference.AsEntity().EntityDefinition();
+            }
+            else if (edmTypeReference.IsComplex())
+            {
+                edmStructuredType = edmTypeReference.AsComplex().ComplexDefinition();
+            }
+            else
+            {
+                throw Error.NotSupported(SRResources.QueryNodeBindingNotSupported, openNode.Kind, typeof(ExpressionBinderBase).Name);
+            }
+
+            return Model.GetDynamicPropertyDictionary(edmStructuredType);
+        }
+
+        /// <summary>
+        /// Analyze previous query and extract grouped properties.
+        /// </summary>
+        /// <param name="source"></param>
+        protected void EnsureFlattenedPropertyContainer(ParameterExpression source)
+        {
+            if (this.BaseQuery != null)
+            {
+                this.HasInstancePropertyContainer = this.BaseQuery.ElementType.IsGenericType
+                    && this.BaseQuery.ElementType.GetGenericTypeDefinition() == typeof(ComputeWrapper<>);
+
+                this.FlattenedPropertyContainer = this.FlattenedPropertyContainer ?? GetFlattenedProperties(source);
             }
         }
 
-        private Expression BindCustomMethodExpressionOrNull(SingleValueFunctionCallNode node)
+        /// <summary>
+        /// Gets expression for property from previously aggregated query
+        /// </summary>
+        /// <param name="propertyPath"></param>
+        /// <returns>Returns null if no aggregations were used so far</returns>
+        protected Expression GetFlattenedPropertyExpression(string propertyPath)
         {
-            Expression[] arguments = BindArguments(node.Parameters);
-            IEnumerable<Type> methodArgumentsType = arguments.Select(argument => argument.Type);
-
-            // Search for custom method info that are binded to the node name
-            MethodInfo methodInfo;
-            if (UriFunctionsBinder.TryGetMethodInfo(node.Name, methodArgumentsType, out methodInfo))
+            if (FlattenedPropertyContainer == null)
             {
-                return MakeFunctionCall(methodInfo, arguments);
+                return null;
             }
 
-            return null;
+            Expression expression;
+            if (FlattenedPropertyContainer.TryGetValue(propertyPath, out expression))
+            {
+                return expression;
+            }
+
+            if (this.HasInstancePropertyContainer)
+            {
+                return null;
+            }
+
+            throw new ODataException(Error.Format(SRResources.PropertyOrPathWasRemovedFromContext, propertyPath));
         }
 
-        private Type RetrieveClrTypeForConstant(IEdmTypeReference edmTypeReference, ref object value)
+        internal string GetFullPropertyPath(SingleValueNode node)
+        {
+            string path = null;
+            SingleValueNode parent = null;
+            switch (node.Kind)
+            {
+                case QueryNodeKind.SingleComplexNode:
+                    var complexNode = (SingleComplexNode)node;
+                    path = complexNode.Property.Name;
+                    parent = complexNode.Source;
+                    break;
+                case QueryNodeKind.SingleValuePropertyAccess:
+                    var propertyNode = ((SingleValuePropertyAccessNode)node);
+                    path = propertyNode.Property.Name;
+                    parent = propertyNode.Source;
+                    break;
+                case QueryNodeKind.SingleNavigationNode:
+                    var navNode = ((SingleNavigationNode)node);
+                    path = navNode.NavigationProperty.Name;
+                    parent = navNode.Source;
+                    break;
+            }
+
+            if (parent != null)
+            {
+                var parentPath = GetFullPropertyPath(parent);
+                if (parentPath != null)
+                {
+                    path = parentPath + "\\" + path;
+                }
+            }
+
+            return path;
+        }
+
+        internal Expression CreatePropertyAccessExpression(Expression source, IEdmProperty property, string propertyPath = null)
+        {
+            string propertyName = Model.GetClrPropertyName(property);
+            propertyPath = propertyPath ?? propertyName;
+            if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True && ExpressionBinderHelper.IsNullable(source.Type) &&
+                source != this.Parameter &&
+                !IsFlatteningSource(source))
+            {
+                Expression cleanSource = ExpressionBinderHelper.RemoveInnerNullPropagation(source, QuerySettings);
+                Expression propertyAccessExpression = null;
+                propertyAccessExpression = GetFlattenedPropertyExpression(propertyPath) ?? Expression.Property(cleanSource, propertyName);
+
+                // source.property => source == null ? null : [CastToNullable]RemoveInnerNullPropagation(source).property
+                // Notice that we are checking if source is null already. so we can safely remove any null checks when doing source.Property
+
+                Expression ifFalse = ExpressionBinderHelper.ToNullable(ConvertNonStandardPrimitives(propertyAccessExpression));
+                return
+                    Expression.Condition(
+                        test: Expression.Equal(source, NullConstant),
+                        ifTrue: Expression.Constant(null, ifFalse.Type),
+                        ifFalse: ifFalse);
+            }
+            else
+            {
+                return GetFlattenedPropertyExpression(propertyPath)
+                    ?? ConvertNonStandardPrimitives(GetPropertyExpression(source, (this.HasInstancePropertyContainer && !propertyPath.Contains("\\", StringComparison.Ordinal) ? "Instance\\" : String.Empty) + propertyName));
+            }
+        }
+
+        internal static Expression GetPropertyExpression(Expression source, string propertyPath)
+        {
+            string[] propertyNameParts = propertyPath.Split('\\');
+            Expression propertyValue = source;
+            foreach (var propertyName in propertyNameParts)
+            {
+                propertyValue = Expression.Property(propertyValue, propertyName);
+            }
+            return propertyValue;
+        }
+
+        // If the expression is of non-standard edm primitive type (like uint), convert the expression to its standard edm type.
+        // Also, note that only expressions generated for ushort, uint and ulong can be understood by linq2sql and EF.
+        // The rest (char, char[], Binary) would cause issues with linq2sql and EF.
+        internal Expression ConvertNonStandardPrimitives(Expression source)
+        {
+            bool isNonstandardEdmPrimitive;
+            Type conversionType = source.Type.IsNonstandardEdmPrimitive(out isNonstandardEdmPrimitive);
+
+            if (isNonstandardEdmPrimitive)
+            {
+                Type sourceType = TypeHelper.GetUnderlyingTypeOrSelf(source.Type);
+
+                Contract.Assert(sourceType != conversionType);
+
+                Expression convertedExpression = null;
+
+                if (TypeHelper.IsEnum(sourceType))
+                {
+                    // we handle enum conversions ourselves
+                    convertedExpression = source;
+                }
+                else
+                {
+                    switch (Type.GetTypeCode(sourceType))
+                    {
+                        case TypeCode.UInt16:
+                        case TypeCode.UInt32:
+                        case TypeCode.UInt64:
+                            convertedExpression = Expression.Convert(ExpressionBinderHelper.ExtractValueFromNullableExpression(source), conversionType);
+                            break;
+
+                        case TypeCode.Char:
+                            convertedExpression = Expression.Call(ExpressionBinderHelper.ExtractValueFromNullableExpression(source), "ToString", typeArguments: null, arguments: null);
+                            break;
+
+                        case TypeCode.DateTime:
+                            convertedExpression = source;
+                            break;
+
+                        case TypeCode.Object:
+                            if (sourceType == typeof(char[]))
+                            {
+                                convertedExpression = Expression.New(typeof(string).GetConstructor(new[] { typeof(char[]) }), source);
+                            }
+                            else if (sourceType == typeof(XElement))
+                            {
+                                convertedExpression = Expression.Call(source, "ToString", typeArguments: null, arguments: null);
+                            }
+#if NETFX // System.Data.Linq.Binary is only supported in the AspNet version.
+                            else if (sourceType == typeof(Binary))
+                            {
+                                convertedExpression = Expression.Call(source, "ToArray", typeArguments: null, arguments: null);
+                            }
+#endif
+                            break;
+
+                        default:
+                            Contract.Assert(false, Error.Format("missing non-standard type support for {0}", sourceType.Name));
+                            break;
+                    }
+                }
+
+                if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True && ExpressionBinderHelper.IsNullable(source.Type))
+                {
+                    // source == null ? null : source
+                    return Expression.Condition(
+                        ExpressionBinderHelper.CheckForNull(source),
+                        ifTrue: Expression.Constant(null, ExpressionBinderHelper.ToNullable(convertedExpression.Type)),
+                        ifFalse: ExpressionBinderHelper.ToNullable(convertedExpression));
+                }
+                else
+                {
+                    return convertedExpression;
+                }
+            }
+
+            return source;
+        }
+
+        internal Expression CreateConvertExpression(ConvertNode convertNode, Expression source)
+        {
+            Type conversionType = Model.GetClrType(convertNode.TypeReference, InternalAssembliesResolver);
+
+            if (conversionType == typeof(bool?) && source.Type == typeof(bool))
+            {
+                // we handle null propagation ourselves. So, if converting from bool to Nullable<bool> ignore.
+                return source;
+            }
+            else if (conversionType == typeof(Date?) &&
+                (source.Type == typeof(DateTimeOffset?) || source.Type == typeof(DateTime?)))
+            {
+                return source;
+            }
+            if ((conversionType == typeof(TimeOfDay?) && source.Type == typeof(TimeOfDay)) ||
+                ((conversionType == typeof(Date?) && source.Type == typeof(Date))))
+            {
+                return source;
+            }
+            else if (conversionType == typeof(TimeOfDay?) &&
+                (source.Type == typeof(DateTimeOffset?) || source.Type == typeof(DateTime?) || source.Type == typeof(TimeSpan?)))
+            {
+                return source;
+            }
+            else if (ExpressionBinderHelper.IsDateAndTimeRelated(conversionType) && ExpressionBinderHelper.IsDateAndTimeRelated(source.Type))
+            {
+                return source;
+            }
+            else if (source == NullConstant)
+            {
+                return source;
+            }
+            else
+            {
+                if (TypeHelper.IsEnum(source.Type))
+                {
+                    // we handle enum conversions ourselves
+                    return source;
+                }
+                else
+                {
+                    // if a cast is from Nullable<T> to Non-Nullable<T> we need to check if source is null
+                    if (QuerySettings.HandleNullPropagation == HandleNullPropagationOption.True
+                        && ExpressionBinderHelper.IsNullable(source.Type) && !ExpressionBinderHelper.IsNullable(conversionType))
+                    {
+                        // source == null ? null : source.Value
+                        return
+                            Expression.Condition(
+                            test: ExpressionBinderHelper.CheckForNull(source),
+                            ifTrue: Expression.Constant(null, ExpressionBinderHelper.ToNullable(conversionType)),
+                            ifFalse: Expression.Convert(ExpressionBinderHelper.ExtractValueFromNullableExpression(source), ExpressionBinderHelper.ToNullable(conversionType)));
+                    }
+                    else
+                    {
+                        return Expression.Convert(source, conversionType);
+                    }
+                }
+            }
+        }
+
+        internal IDictionary<string, Expression> GetFlattenedProperties(ParameterExpression source)
+        {
+            if (this.BaseQuery == null)
+            {
+                return null;
+            }
+
+            if (!typeof(GroupByWrapper).IsAssignableFrom(BaseQuery.ElementType))
+            {
+                return null;
+            }
+
+            var expression = BaseQuery.Expression as MethodCallExpression;
+            if (expression == null)
+            {
+                return null;
+            }
+
+            // After $apply we could have other clauses, like $filter, $orderby etc.
+            // Skip of filter expressions
+            expression = SkipFilters(expression);
+
+            if (expression == null)
+            {
+                return null;
+            }
+
+            var result = new Dictionary<string, Expression>();
+            CollectContainerAssignments(source, expression, result);
+            if (this.HasInstancePropertyContainer)
+            {
+                var instanceProperty = Expression.Property(source, "Instance");
+                if (typeof(DynamicTypeWrapper).IsAssignableFrom(instanceProperty.Type))
+                {
+                    var computeExpression = expression.Arguments.FirstOrDefault() as MethodCallExpression;
+                    computeExpression = SkipFilters(computeExpression);
+                    if (computeExpression != null)
+                    {
+                        CollectContainerAssignments(instanceProperty, computeExpression, result);
+                    }
+                }
+            }
+
+            return result;
+        }
+
+        internal Type RetrieveClrTypeForConstant(IEdmTypeReference edmTypeReference, ref object value)
         {
             Type constantType = Model.GetClrType(edmTypeReference, InternalAssembliesResolver);
 
@@ -1978,6 +1305,40 @@ namespace Microsoft.AspNetCore.OData.Query.Expressions
             }
 
             return constantType;
+        }
+
+        internal Expression BindCastToEnumType(Type sourceType, Type targetClrType, QueryNode firstParameter, int parameterLength)
+        {
+            Type enumType = TypeHelper.GetUnderlyingTypeOrSelf(targetClrType);
+            ConstantNode sourceNode = firstParameter as ConstantNode;
+
+            if (parameterLength == 1 || sourceNode == null || sourceType != typeof(string))
+            {
+                // We only support to cast Enumeration type from constant string now,
+                // because LINQ to Entities does not recognize the method Enum.TryParse.
+                return NullConstant;
+            }
+            else
+            {
+                object[] parameters = new[] { sourceNode.Value, Enum.ToObject(enumType, 0) };
+                bool isSuccessful = (bool)EnumTryParseMethod.MakeGenericMethod(enumType).Invoke(null, parameters);
+
+                if (isSuccessful)
+                {
+                    if (QuerySettings.EnableConstantParameterization)
+                    {
+                        return LinqParameterContainer.Parameterize(targetClrType, parameters[1]);
+                    }
+                    else
+                    {
+                        return Expression.Constant(parameters[1], targetClrType);
+                    }
+                }
+                else
+                {
+                    return NullConstant;
+                }
+            }
         }
     }
 }
